added expected exceptions decorator, some fixes based on @painassasin comments

+++ project/services/favorite_service.py
@@ -1,4 +1,5 @@
 from project.dao.main_dao import FavoriteDAO
+from project.exceptions import ItemNotFound
 
 
 class FavoriteService:
@@ -10,7 +11,7 @@ class FavoriteService:
         return new_favorite
 
     def delete_favorite(self, user_id: int, movie_id: int) -> dict:
-        favorite_to_delete_id: int = self.dao.select_unique_item_id_by_arguments(user_id=user_id, movie_id=movie_id)
-        favorite_to_delete: dict = self.dao.delete_item_by_pk(favorite_to_delete_id)
-
-        return favorite_to_delete
+        if favorite_to_delete_id := self.dao.select_unique_item_id_by_arguments(user_id=user_id, movie_id=movie_id):
+            favorite_to_delete: dict = self.dao.delete_item_by_pk(favorite_to_delete_id)
+            return favorite_to_delete
+        raise ItemNotFound(f'Favorite with user_id={user_id} and movie_id={movie_id} not exists.')
